Extract the common code-block to a seperate file
